refactor: Extracted closure functions for report generator into helpers module [#1094]
+ Refactored closures to import from helpers
