+++ django_api/django_api/apps/core/helpers.py
@@ -1,5 +1,6 @@
 import math
 import random
+import logging
 import sys
 import os
 from ast import literal_eval
@@ -12,6 +13,17 @@ from itertools import combinations, product
 from core.common import (
     PD_FREQUENCY_LEVEL,
 )
+
+from indicator.models import (
+    IndicatorBlueprint,
+    IndicatorReport,
+    IndicatorLocationData,
+    ReportingEntity,
+    Reportable,
+)
+from unicef.models import ProgressReport
+
+logger = logging.getLogger(__name__)
 
 
 @contextmanager
@@ -338,3 +350,269 @@ def find_missing_frequency_period_dates_for_indicator_report(indicator, latest_i
                 date_list.append(missing_date)
 
     return date_list
+
+
+def get_latest_pr_by_type(pd, report_type):
+    """
+    Return latest ProgressReport instance given report_type
+
+    Arguments:
+        report_type {str} -- A report type as string: [QPR, HR, SR]
+
+    Returns:
+        ProgressReport -- Latest ProgressReport instance for given report_type
+    """
+
+    if report_type == "QPR":
+        return pd.progress_reports \
+            .filter(report_type="QPR").order_by('start_date').last()
+
+    if report_type == "HR":
+        return pd.progress_reports \
+            .filter(report_type="HR").order_by('start_date').last()
+
+    if report_type == "SR":
+        return pd.progress_reports \
+            .filter(report_type="SR").order_by('due_date').last()
+
+
+def create_pr_for_report_type(pd, idx, reporting_period, generate_from_date):
+    """
+    Create ProgressReport instance by its ReportingPeriodDate instance's report type
+
+    Arguments:
+        pd {ProgrammeDocument} -- ProgrammeDocument instance for ProgressReport to generate
+        idx {int} -- Integer to denote report number
+        reporting_period {ReportingPeriodDates} -- ReportingPeriodDates instance for new ProgressReport
+        generate_from_date {datetime.datetime} -- datetime instance from latest ProgressReport on same report
+
+    Returns:
+        Tuple[ProgressReport, datetime.datetime, datetime.datetime, datetime.datetime]
+        - Newly generated ProgressReport & 3 datetime objects
+    """
+
+    end_date = reporting_period.end_date
+    due_date = reporting_period.due_date
+    start_date = reporting_period.start_date
+
+    # Create ProgressReport first
+    logger.info("Creating ProgressReport for {} - {}".format(start_date, end_date))
+
+    # Re-query latest ProgressReport by report type
+    latest_progress_report = get_latest_pr_by_type(pd, reporting_period.report_type)
+
+    if latest_progress_report:
+        report_type = latest_progress_report.report_type
+        report_number = latest_progress_report.report_number + 1
+        is_final = idx == pd.reporting_periods.filter(report_type=reporting_period.report_type).count() - 1
+
+    else:
+        report_number = 1
+        report_type = reporting_period.report_type
+        is_final = False
+
+    next_progress_report = ProgressReport.objects.create(
+        start_date=start_date,
+        end_date=end_date,
+        due_date=due_date,
+        programme_document=pd,
+        report_type=report_type,
+        report_number=report_number,
+        is_final=is_final,
+    )
+
+    return (next_progress_report, start_date, end_date, due_date)
+
+
+def create_pr_ir_for_reportable(pd, reportable, pai_ir_for_period, start_date, end_date, due_date):
+    if reportable.blueprint.unit == IndicatorBlueprint.NUMBER:
+        logger.info("Creating Quantity IndicatorReport for {} - {}".format(start_date, end_date))
+        indicator_report = IndicatorReport.objects.create(
+            progress_report=None,
+            reportable=reportable,
+            parent=pai_ir_for_period,
+            time_period_start=start_date,
+            time_period_end=end_date,
+            due_date=due_date,
+            title=reportable.blueprint.title,
+            total={'c': 0, 'd': 0, 'v': 0},
+            overall_status="NoS",
+            report_status="Due",
+            submission_date=None,
+            reporting_entity=ReportingEntity.objects.get(title="UNICEF"),
+        )
+
+        for location_goal in reportable.reportablelocationgoal_set.filter(is_active=True):
+            logger.info("Creating IndicatorReport {} IndicatorLocationData for {} - {}".format(
+                indicator_report, start_date, end_date
+            ))
+            IndicatorLocationData.objects.create(
+                indicator_report=indicator_report,
+                location=location_goal.location,
+                num_disaggregation=indicator_report.disaggregations.count(),
+                level_reported=indicator_report.disaggregations.count(),
+                disaggregation_reported_on=list(indicator_report.disaggregations.values_list(
+                    'id', flat=True)),
+                disaggregation={
+                    '()': {'c': 0, 'd': 0, 'v': 0}
+                },
+            )
+
+    else:
+        logger.info("Creating PD {} Ratio IndicatorReport for {} - {}".format(pd, start_date, end_date))
+        indicator_report = IndicatorReport.objects.create(
+            progress_report=None,
+            reportable=reportable,
+            parent=pai_ir_for_period,
+            time_period_start=start_date,
+            time_period_end=end_date,
+            due_date=due_date,
+            title=reportable.blueprint.title,
+            total={'c': 0, 'd': 0, 'v': 0},
+            overall_status="NoS",
+            report_status="Due",
+            submission_date=None,
+            reporting_entity=ReportingEntity.objects.get(title="UNICEF"),
+        )
+
+        for location_goal in reportable.reportablelocationgoal_set.filter(is_active=True):
+            logger.info("Creating IndicatorReport {} IndicatorLocationData {} - {}".format(
+                indicator_report, start_date, end_date
+            ))
+            IndicatorLocationData.objects.create(
+                indicator_report=indicator_report,
+                location=location_goal.location,
+                num_disaggregation=indicator_report.disaggregations.count(),
+                level_reported=indicator_report.disaggregations.count(),
+                disaggregation_reported_on=list(indicator_report.disaggregations.values_list(
+                    'id', flat=True)),
+                disaggregation={
+                    '()': {'c': 0, 'd': 0, 'v': 0}
+                },
+            )
+
+    return indicator_report
+
+
+def create_ir_and_ilds_for_pr(pd, reportable_queryset, next_progress_report, start_date, end_date, due_date):
+    """
+    Create a set of new IndicatorReports and IndicatorLocationData instances per
+    IndicatorReport instance, with passed-in new dates and new ProgressReport instance
+    Arguments:
+        pd {ProgrammeDocument} -- ProgrammeDocument instnace
+        reportable_queryset {django.Queryset[Reportable]} -- Reportable queryset on LLO
+        next_progress_report {ProgressReport} -- Newly generated Progress Report instance
+        start_date {datetime.datetime} -- Start date for reporting
+        end_date {datetime.datetime} -- End date for reporting
+        due_date {datetime.datetime} -- due date for reporting
+    """
+
+    if next_progress_report.report_type != "SR":
+        if next_progress_report.report_type == "QPR":
+            queryset = reportable_queryset
+
+        else:
+            # Filter non-Cluster reportables first
+            queryset = reportable_queryset.filter(
+                ca_indicator_used_by_reporting_entity__isnull=True,
+                is_unicef_hf_indicator=True
+            )
+
+        ir_list = list()
+
+        for reportable in queryset:
+            indicator_report = create_pr_ir_for_reportable(
+                reportable,
+                None,
+                start_date,
+                end_date,
+                due_date,
+            )
+            indicator_report.progress_report = next_progress_report
+            indicator_report.save()
+            ir_list.append(indicator_report)
+
+        if next_progress_report.report_type == "HR":
+            hr_reports = list()
+
+            # If there are no UNICEF HF indicator reports then delete blank ProgressReport
+            if len(ir_list) == 0:
+                # Re-assign report_number to new HR
+                report_number = next_progress_report.report_number
+                next_progress_report.delete()
+
+            else:
+                # Pre-populate new HR report_number in case a new Progress Report needs to be generated
+                report_number = next_progress_report.report_number + 1
+                hr_reports.append(next_progress_report)
+
+            # Process cluster Reportables separately
+            for reportable in reportable_queryset.filter(ca_indicator_used_by_reporting_entity__isnull=False):
+                cai_indicator = reportable.ca_indicator_used_by_reporting_entity
+                pai_ir_for_period = None
+
+                # If LLO indicator has ClusterActivity Indicator ID reference,
+                # find the adopted PartnerActivity indicator from ClusterActivity Indicator
+                # with LLO's Partner ID
+                # and grab a corresponding IndicatorReport from ClusterActivity Indicator
+                # given the start & end date
+                if cai_indicator:
+                    try:
+                        # Grabbing first adopted partner activity in case
+                        # multiple adopted partner activities happen, although this is illegal state!
+                        pai_indicator = cai_indicator.children \
+                            .filter(partner_activities__partner=pd.partner) \
+                            .first()
+                        pai_ir_for_period = pai_indicator.indicator_reports.get(
+                            time_period_start=start_date,
+                            time_period_end=end_date,
+                        )
+
+                        if pai_ir_for_period:
+                            indicator_report = create_pr_ir_for_reportable(
+                                reportable,
+                                pai_ir_for_period,
+                                pai_ir_for_period.time_period_start,
+                                pai_ir_for_period.time_period_end,
+                                pai_ir_for_period.due_date,
+                            )
+
+                            # Bundle this cluster LLO Indicator report to HR progress report generated so far
+                            # for this iteration if the dates are matching
+                            for hr_report in hr_reports:
+                                if indicator_report.time_period_start == hr_report.start_date \
+                                        and indicator_report.time_period_end == hr_report.end_date \
+                                        and indicator_report.due_date == hr_report.due_date:
+                                    indicator_report.progress_report = hr_report
+                                    break
+
+                            if not indicator_report.progress_report:
+                                # Otherwise, create a brand new HR progress report
+                                # for this cluster LLO Indicator report
+                                new_cluster_hr_progress_report = ProgressReport.objects.create(
+                                    start_date=indicator_report.time_period_start,
+                                    end_date=indicator_report.time_period_end,
+                                    due_date=indicator_report.due_date,
+                                    programme_document=pd,
+                                    report_type="HR",
+                                    report_number=report_number,
+                                    is_final=False,
+                                )
+                                indicator_report.progress_report = new_cluster_hr_progress_report
+
+                                # Increment report_number for next HR progress report to be created if needed
+                                report_number += 1
+                                hr_reports.append(new_cluster_hr_progress_report)
+
+                            indicator_report.save()
+
+                    except Reportable.DoesNotExist as e:
+                        logger.exception(
+                            "FAILURE: CANNOT FIND adopted PartnerActivity Reportable "
+                            "from given ClusterActivity Reportable and PD Partner ID. "
+                            "Skipping link!", e)
+                    except IndicatorReport.DoesNotExist as e:
+                        logger.exception(
+                            "FAILURE: CANNOT FIND IndicatorReport from adopted PartnerActivity Reportable "
+                            "linked with LLO Reportable. "
+                            "Skipping link!", e)
